AppCode/requirements.txt
ipython
numpy
pandas
psycopg2_binary
PyPDF2
scikit_learn
sentence_transformers
skillNer
spacy
en-core-web-lg @https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
streamlit